:green_heart: [#7] specテストを追加
