+++ Gemfile
@@ -74,4 +74,5 @@ group :test do
   gem "selenium-webdriver"
   gem "webdrivers"
   gem "shoulda-matchers"
+  gem "rails-controller-testing"
 end
